+++ app/api/types.d.ts
@@ -4,4 +4,20 @@ export interface I_MakeRes {
     severity: T_Severity;
     message: string;
     data?: any;
+}
+
+export type T_Email = {
+
+    from: {
+        label: string;
+        email: string;
+    },
+    to: {
+        label: string;
+        email: string;
+    }
+    subject: string;
+    body: string; // Markdown or HTML content?  
+    template?: string;
+
 };
